Cost value changed from double to Integer

+++ API/src/main/java/com/maimai/tamagotchi/item/DefaultType.java
@@ -6,5 +6,5 @@ public interface DefaultType {
 
     double getValue();
 
-    double getCost();
+    int getCost();
 }
